Add cmime_base64_decode_buffer.
Currently, the base64 decoding seems broken in significant ways.

Firstly, an attachment might have 0 bytes in it (bytes with a
value 0, not being 0 bytes long) - for example, the base64
encoded PDF in the email I am testing.

Decoding the block and then using strlen therefore returns a
truncated result. Accordingly, introduce cmime_base64_decode_buffer
that does the same as cmime_base64_decode_string, but also
returns the byte length of the result.

Having fixed this, it became clear that my base64 encoded PDF
was decoding badly because the decoding code didn't cope with
the newlines (actually CRLF) in the mime encoding. And nowhere
did the code attempt to cope with padding markers at the end
of the code.

Simply spotting undefined chars and interpreting them as 0 does
not help here. So in this new version, out of range chars are
ignored.